Problem statement
You are to create a screen in React Native for iOS. This screen contains a list of items where each item displays a number and a random string (to be explained later). Each list item will also have some random amount of empty space below the text it displays (between 10 and 50px). The numbers in each item are to be in order (0 at the bottom of the list). When the app is opened, you are to show 50 items (0 - 49 - so the app should start from the list at the bottom) in the list and whenever the user scrolls all the way to the top of the list, you need to show an activity indicator (for 800 MS) and then the list is to add 50 more items on top. The scroll position should not change during this addition.
There is also a button on top of the screen (not top of the list) and when it is clicked, the list is supposed to scroll down such that item 0 is visible at the bottom of the screen.

The random string in each list item has two words. Each word should have a length of 3 to 15 characters and have only lowercase characters (a-z). On top of the list, there is also a search bar which the user can use to search the strings in the list. For example, if the list contains the string [“dlsf kodsfh”, “dad asdw”, “sfaf ruvta”] and the user searches for “d”, the result should be [“dlsf kodsfh”, “dad asdw”]. If the user searches for “a”, then the list should be [“dad asdw”] and so on… The list should show the result as the user types each letter. Please balance between time and space complexity here.

NOTE: You are to start at the bottom of the list when the app opens, i.e. number 0 should be visible at the bottom of the screen when the app starts

The designing of this screen is up to you.

Please use Javascript ES7.